Makes the 'val' method set the query silently.
Currently, the only time 'silent' is used when setting the query is in
the default 'select' method; an item is selected, and the value is set
silently such that another query to the autocompleter won't be made. I
believe this should always happen when triggering $element.typeahead('val',
aValue);

When I manually call 'val' on a typeahead object, it is never because I
want to autocomplete a result. It is because I am populating that value
either from a selected dataset (custom render/select method) or from
another source. This redundant querying can cause problems when the
source is a remote that charges per request, for example (mapquest
autocomplete addresses, for one).

If you *truly* want to fill in an input value and trigger an
autocomplete request, use the new 'query' method instead.